+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-# 0.1.0 / _Not released yet_
+# 1.0.0 / 2024-09-21
 
-* First release!
+- Drop support for 1password CLI v1.x.
+- Use `op inject` instead of `op run` to make Ctrl+C work.
+- Add option `--no-overwrite` to avoid overwriting already set environment variables.
+- Add option `--verbose` for more output while loading direnv.
+
+# 0.1.0 / 2022-01-18
+
+- First release!